refactor: auto-split test_grip_contact_model.py
